ISSUE-1545 Move BlobCalendarResolver to jmap-extensions-api
